ConstructorTabsTestParams renamed to ConstructorTabsParamsTest
